requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0